Convert code to PlatformIO
Convert source code structure. configurations and settings so that they work with PlatformIO
