+++ README.md
@@ -93,7 +93,7 @@ Note that the battery terminals nor the power switch are currently included in t
 The battery holder might be a bit tight, so some sanding/cutting of the edges may be necessary.
  
 ## Flashing the firmware
-The firmware was built with Arduino IDE v 2.3.2. Using the Board Manager, one needs to install esp32 by Espressif (version 3.0.2).
+The firmware was built with PlatformIO v core 6.1.18, home 3.4.4. One needs to install lilygo-t-display board (or it may get installed automatically when building...)
 
 Following external libraries are needed:
 
@@ -104,15 +104,14 @@ Other drivers are included in the repository with necessary modifications.
 
 To build the software select following as the board settings:
 
- - Board: ESP32 Dev Module
- - Upload Speed: 921600
- - CPU Frequency: 240Mhz (WiFi/BT)
- - Flash Frequency: 80Mhz
- - Flash Mode: QIO
- - Flash Size: 4MB (32Mb)
- - Partition Scheme: Default 4MB with spiffs (1.2MB APP/1.5 SPIFFS)
- - Core Debug Level: None --> For debugging can set also different values, for release select None
- - PSRAM: Disabled
+ * Board: Lilygo-t-display
+ * Upload Speed: 921600
+ * CPU Frequency: 240Mhz (WiFi/BT)
+ * Flash Frequency: 80Mhz
+ * Flash Mode: QIO
+ * Partition Scheme: Default.csv [4MB with spiffs (1.2MB APP/1.5 SPIFFS)]
+ * Core Debug Level: None --> For debugging can set also different values, for release select None
+ * PSRAM: No
 
  
 # Disclaimer
@@ -121,7 +120,7 @@ to be used for reference. Everything is provided as is, without any warranty. Th
 liable for any damages or other liability arising from use of anything provided here.
 
 # Copyright
-Copyright 2024 Lauri Peltonen
+Copyright 2024, 2025 Lauri Peltonen
 
 Licensed under GPLv3.
