+++ Src-PlatformIO/.gitignore
@@ -1,0 +1,24 @@
+.pio
+.vscode/.browse.c_cpp.db*
+.vscode/c_cpp_properties.json
+.vscode/launch.json
+.vscode/ipch
+
+# PlatformIO build directory
+.pio/
+
+# VS Code settings
+.vscode/
+
+# Compiled binaries and temporary files
+*.elf
+*.bin
+*.hex
+*.map
+
+# Operating system files
+.DS_Store
+Thumbs.db
+
+# Release artifacts
+releases/
