+++ Src-PlatformIO/platformio.ini
@@ -1,0 +1,30 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+
+[env:lilygo-t-display]
+build_type = debug
+build_flags = -DCORE_DEBUG_LEVEL=4
+platform = espressif32
+board = lilygo-t-display
+framework = arduino
+lib_deps = 
+	bodmer/TFT_eSPI@^2.5.43
+	h2zero/NimBLE-Arduino@1.4.1
+upload_speed = 921600
+board_build.f_cpu = 240000000L
+board_build.f_flash = 80000000L
+board_build.flash_mode = qio
+board_build.partitions = default.csv
+;board_build.partitions = huge_app.csv
+monitor_speed = 115200
+monitor_filters = esp32_exception_decoder
+
+[platformio]
+description = DIY portable spirometer for measuring VO2max
