--- Src/status.h
@@ -1,15 +1,0 @@
-/*
- * Copyright (C) 2024 Lauri Peltonen
- * GPL V3
- */
-
-#ifndef __STATUS_H__
-#define __STATUS_H__
-
-void initScreens();
-void showScreen();
-void changeScreen(bool forwards);
-void setScreen(int n);
-void statusInitial();
-
-#endif
